backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
requests
motor
python-multipart
pathlib
torch
transformers
speechrecognition
PyJWT
passlib
bcrypt
sqlalchemy 
python-jose
python-dotenv
pymongo
huggingface_hub